pipelines/content-generation/prizm-ai-v5/pyproject.toml
[project]
name = "prizm-ai"
version = "5.0.0"
description = "Système de génération automatique d'articles B2B sur l'IA pour PME françaises"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Samuel", email = "contact@prizm-ai.com"}
]
keywords = ["ai", "content-generation", "langgraph", "b2b", "automation"]

dependencies = [
    # Framework agents
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    
    # Évaluation qualité
    "ragas>=0.2.0",
    "datasets>=2.0.0",
    
    # APIs externes
    "openai>=1.0.0",
    "httpx>=0.27.0",
    
    # Traitement texte et données
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tiktoken>=0.7.0",
    
    # Utilitaires
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "aiofiles>=24.0.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    
    # Observabilité
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
prizm = "prizm_ai.main:main"
prizm-test = "prizm_ai.main:test_run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prizm_ai"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
